@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.13.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -37,6 +37,7 @@ dependencies:
   universal_html: ^2.0.8
   # firebase_storage_web: ^3.0.3
   xml: ^5.3.0
+  flutter_login: ^4.1.0
 
 dev_dependencies:
   flutter_test:
